requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
django-celery-results==2.5.1
boto3==1.34.144
django-storages==1.14.4
python-dotenv==1.0.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn==0.30.1
# Bio analysis dependencies
scanpy==1.10.2
anndata==0.10.8
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2